Keep reference to original response object in error cases
Sometimes, you need to handle error cases differently depending on what
the actual error was.  For example, zipcode errors would be handled
differently than missing line items.  By keeping a reference to the
original object, it's easier to programatically parse and understand the
error message rather than trying to trying to check whether some
substring is in the error message.

This change should be 100% backwards compatible but provides an
enhancement for future users.